Fix XKB compilation
The error was:

error[E0308]: mismatched types
    --> x11rb-protocol/src/protocol/xkb.rs:6663:13
     |
6663 |             clear,
     |             ^^^^^ expected struct `EventType`, found `u16`

error[E0308]: mismatched types
    --> x11rb-protocol/src/protocol/xkb.rs:6664:13
     |
6664 |             select_all,
     |             ^^^^^^^^^^ expected struct `EventType`, found `u16`

error[E0308]: mismatched types
    --> x11rb-protocol/src/protocol/xkb.rs:8628:13
     |
8628 |             virtual_mods,
     |             ^^^^^^^^^^^^ expected struct `VMod`, found `u16`

error[E0308]: mismatched types
     --> x11rb-protocol/src/protocol/xkb.rs:10555:13
      |
10555 |             virtual_mods,
      |             ^^^^^^^^^^^^ expected struct `VMod`, found `u16`

error[E0308]: mismatched types
     --> x11rb-protocol/src/protocol/xkb.rs:10561:13
      |
10561 |             group_names,
      |             ^^^^^^^^^^^ expected struct `SetOfGroup`, found `u8`

For more information about this error, try `rustc --explain E0308`.
error: could not compile `x11rb-protocol` due to 5 previous errors

The problem was that there was a call to into(), but it figured out the
wrong type due to a later function call to a function that expected u16.
Fix this by introducing another conversion in the call to that function.

The code in question looks like this:

        let (select_all, remaining) = u16::try_parse(remaining)?;
        let select_all = select_all.into();
        [...]
        let (details, remaining) = SelectEventsAux::try_parse(remaining, [...] u16::from(select_all))?;
        let _ = remaining;
        Ok(SelectEventsRequest {
            device_spec,
            clear,
            select_all,
            affect_map,
            map,
            details: Cow::Owned(details),
        })

The try_parse() function that is called here expected an u16 argument thus the
select_all.into() before figured that type out. The SelectEventsRequests
construction at the end thus failed.

+++ x11rb-protocol/src/protocol/render.rs
@@ -2050,7 +2050,7 @@ impl<'input> CreatePictureRequest<'input> {
         let (drawable, remaining) = xproto::Drawable::try_parse(remaining)?;
         let (format, remaining) = Pictformat::try_parse(remaining)?;
         let (value_mask, remaining) = u32::try_parse(remaining)?;
-        let (value_list, remaining) = CreatePictureAux::try_parse(remaining, value_mask)?;
+        let (value_list, remaining) = CreatePictureAux::try_parse(remaining, u32::from(value_mask))?;
         let _ = remaining;
         Ok(CreatePictureRequest {
             pid,
@@ -2442,7 +2442,7 @@ impl<'input> ChangePictureRequest<'input> {
         }
         let (picture, remaining) = Picture::try_parse(value)?;
         let (value_mask, remaining) = u32::try_parse(remaining)?;
-        let (value_list, remaining) = ChangePictureAux::try_parse(remaining, value_mask)?;
+        let (value_list, remaining) = ChangePictureAux::try_parse(remaining, u32::from(value_mask))?;
         let _ = remaining;
         Ok(ChangePictureRequest {
             picture,
